[refactor] renderDaily와 관련된 함수 리팩토링 - main태그에 appendChild 되는 과정을 renderSelectDayDaily 라는 함수로 분리시켰다.
