+++ JS/render.js
@@ -1,7 +1,6 @@
 // ============== renderMain 관련 데이터 ==============
 import {containerInfo} from './data/containerInfoData.js'
 import {todayWebtoonsData} from './data/todayWebtoons.js'
-import {dailyTopData} from './data/scrape/dailyTopData.js'
 import mainBannerData from './data/json/mainBannerData.json' assert { type: "json"} ;
 // ============== renderHome 관련 모듈 ==============
 import {renderContainer} from './components/container.js'
@@ -9,10 +8,7 @@ import {renderMainBanner} from './components/banner/bannerMain.js'
 import {renderPromotionBanner} from './components/banner/bannerPromotion.js'
 import {renderMoveApp} from './components/moveApp.js'
 // ============== renderDaily 관련 모듈 ==============
-import {makeSelectDayDailySection} from './components/selectDay/selectDayDaily.js'
-import {makeWebtoonList} from './components/webtoonList.js'
-
-import {$, $all} from './utility.js'
+import {renderSelectDayDaily} from './components/selectDay/selectDayDaily.js'
 
 function renderMain(tab) {
   const tabInfo = {
@@ -23,7 +19,7 @@ function renderMain(tab) {
     },
     daily: {
       func: renderDaily,
-      data: dailyTopData,
+      data: todayWebtoonsData,
     }
   }
 
@@ -43,13 +39,7 @@ function renderHome(tab, data, layoutInfo) {
 
 function renderDaily(tab, data) {
   renderMainBanner(mainBannerData[tab]);
-  const main = $('.main');
-  const newList = makeSelectDayDailySection();
-  main.appendChild(newList)
-
-  const DAILY_ITEMS = 10;
-  const webtoonsList = makeWebtoonList(DAILY_ITEMS, data);
-  newList.appendChild(webtoonsList);
+  renderSelectDayDaily(data);
 }
 
 export {renderMain}
